Factor out a package.team mixin and smooth it out
